+++ Changelog.md
@@ -4,6 +4,7 @@
 
 ## [v2.2.1]
 - Fix bug where error was raised when viewing jupyter notebooks if python is not available (#6418)
+- Apply course maximum file size to feedback files (#6430)
 - Fix bug where peer reviewers could not view results they are assigned to review (#6439)
 - Make fewer connections when connecting to redis (#6443)